Show the organization's workload states on its Overview
The workload state bar extracts to WorkloadStateCard — the Dashboard and
the org Overview are its two consumers — with an optional organizationId
filter and the same in-place refetch on org switch. The org Overview
places it beside the Details panel.
