+++ kinotic-frontend/apps/system/src/pages/Dashboard.vue
@@ -30,37 +30,7 @@
         </div>
       </div>
 
-      <div class="rounded-lg border border-surface p-4">
-        <div class="flex items-start justify-between">
-          <div>
-            <h2 class="text-base font-semibold">Workloads</h2>
-            <p class="mb-4 text-xs text-muted-color">
-              Every workload on the platform, by state.
-            </p>
-          </div>
-          <RouterLink to="/worker-nodes" class="text-sm text-muted-color hover:text-color">View all</RouterLink>
-        </div>
-        <div v-if="workloadTotal === 0" class="py-6 text-center text-sm text-muted-color">
-          No workloads
-        </div>
-        <template v-else>
-          <div class="flex h-3 gap-[2px] overflow-hidden rounded">
-            <div
-              v-for="seg in workloadSegments.filter(s => s.count > 0)"
-              :key="seg.label"
-              :style="{ width: (seg.count / workloadTotal * 100) + '%', background: seg.color }"
-              :title="`${seg.label}: ${seg.count}`"
-            />
-          </div>
-          <div class="mt-3 flex flex-wrap gap-x-4 gap-y-1">
-            <div v-for="seg in workloadSegments" :key="seg.label" class="flex items-center gap-1.5 text-sm">
-              <span class="h-2.5 w-2.5 rounded-full" :style="{ background: seg.color }" />
-              <span class="text-muted-color">{{ seg.label }}</span>
-              <span class="font-medium">{{ seg.count }}</span>
-            </div>
-          </div>
-        </template>
-      </div>
+      <WorkloadStateCard description="Every workload on the platform, by state." view-all-to="/worker-nodes" />
     </div>
 
     <div class="mt-6 rounded-lg border border-surface">
@@ -108,17 +78,17 @@
 
 <script setup lang="ts">
 import { computed, onMounted, ref } from 'vue'
-import { RouterLink } from 'vue-router'
 import Button from 'primevue/button'
 import Message from 'primevue/message'
 import Tag from 'primevue/tag'
 
 import { Kinotic, Pageable } from '@kinotic-ai/core'
-import { WorkloadStatus, type KinoticClusterInfo, type VmNode } from '@kinotic-ai/os-api'
+import type { KinoticClusterInfo, VmNode } from '@kinotic-ai/os-api'
 
 import { PageHeader, formatMb, isDark } from '@kinotic-ai/frontend-common'
 
 import LogLevelDialog from '@/components/LogLevelDialog.vue'
+import WorkloadStateCard from '@/components/WorkloadStateCard.vue'
 import StatTile, { type StatTileAccent } from '@/components/StatTile.vue'
 
 const clusterInfo = ref<KinoticClusterInfo | null>(null)
@@ -126,39 +96,8 @@ const clusterError = ref<string | null>(null)
 const organizationCount = ref<number | null>(null)
 const workerNodeCount = ref<number | null>(null)
 const workerNodes = ref<VmNode[]>([])
-const workloadCounts = ref<Record<StatusBucket, number>>({ running: 0, starting: 0, stopping: 0, stopped: 0, failed: 0 })
-
-type StatusBucket = 'running' | 'starting' | 'stopping' | 'stopped' | 'failed'
-
-// Theme ramp steps validated for adjacent-pair CVD separation and surface contrast in both
-// modes (dark uses the lighter 400 steps). STOPPED is a deliberate achromatic neutral, and
-// every segment carries a labeled legend row, so identity is never color alone.
-const STATUS_SERIES: { bucket: StatusBucket; label: string; light: string; dark: string }[] = [
-  { bucket: 'running', label: 'Running', light: '#22C55E', dark: '#4ADE80' },
-  { bucket: 'starting', label: 'Starting', light: '#0EA5E9', dark: '#38BDF8' },
-  { bucket: 'stopping', label: 'Stopping', light: '#F97316', dark: '#FB923C' },
-  { bucket: 'stopped', label: 'Stopped', light: '#6B7280', dark: '#9CA3AF' },
-  { bucket: 'failed', label: 'Failed', light: '#EF4444', dark: '#F87171' }
-]
-
-const BUCKET_BY_STATUS: Record<WorkloadStatus, StatusBucket> = {
-  [WorkloadStatus.RUNNING]: 'running',
-  [WorkloadStatus.PENDING]: 'starting',
-  [WorkloadStatus.STARTING]: 'starting',
-  [WorkloadStatus.STOPPING]: 'stopping',
-  [WorkloadStatus.STOPPED]: 'stopped',
-  [WorkloadStatus.FAILED]: 'failed'
-}
 
 const capacityColor = computed(() => isDark.value ? '#38BDF8' : '#0EA5E9')
-
-const workloadSegments = computed(() => STATUS_SERIES.map(series => ({
-  label: series.label,
-  count: workloadCounts.value[series.bucket],
-  color: isDark.value ? series.dark : series.light
-})))
-
-const workloadTotal = computed(() => workloadSegments.value.reduce((sum, seg) => sum + seg.count, 0))
 
 const capacityRows = computed(() => {
   const total = { cpus: 0, memoryMb: 0, diskMb: 0 }
@@ -258,28 +197,5 @@ onMounted(async () => {
   } catch {
     // Same em-dash fallback as the organization count; the capacity card shows its empty state
   }
-  try {
-    await loadWorkloadCounts()
-  } catch {
-    // The workloads card shows its empty state
-  }
 })
-
-async function loadWorkloadCounts() {
-  const counts: Record<StatusBucket, number> = { running: 0, starting: 0, stopping: 0, stopped: 0, failed: 0 }
-  // Buckets are counted client-side from pages; the loop is bounded, so a platform with
-  // more than 1000 workloads undercounts — revisit with server-side aggregation then
-  const pageSize = 100
-  for (let pageNumber = 0; pageNumber < 10; pageNumber++) {
-    const page = await Kinotic.workloads.findAll(Pageable.create(pageNumber, pageSize))
-    const content = page.content ?? []
-    for (const workload of content) {
-      counts[BUCKET_BY_STATUS[workload.status]] += 1
-    }
-    if (content.length < pageSize) {
-      break
-    }
-  }
-  workloadCounts.value = counts
-}
 </script>
